requirements.txt
torch
torchvision
transformers
accelerate
peft
datasets
pillow
imageio
requests
azure-storage-blob
numpy
pandas
